pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aeda-ai"
version = "0.1.0"
description = "Automated Environmental Exploratory Data Analysis framework for LEA-CEAC"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "plotly>=5.15",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
umap = ["umap-learn>=0.5"]
shap = ["shap>=0.42"]
ui = ["streamlit>=1.28"]
all = ["umap-learn>=0.5", "shap>=0.42", "streamlit>=1.28"]
dev = ["pytest>=7.4", "pytest-cov>=4.1"]

[tool.setuptools.packages.find]
include = ["aeda*"]
